requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2-binary
django-debug-toolbar
django-debug-toolbar-force
bpython
social-auth-app-django
django-filter
django-environ
